feat!: add Spacer; remove FixedSpace
